# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Morex API",
        "description": "API pour l'application mobile Morex - Gestion des finances personnelles",
        "contact": {
            "name": "Morex Support",
            "email": "support@morex.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/debts": {
            "get": {
                "tags": [
                    "Debts"
                ],
                "summary": "Liste des dettes et créances",
                "operationId": "a04e56aa752df98fea015c614c68f30d",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "debt = je dois, credit = on me doit",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "debt",
                                "credit"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "paid",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des dettes/créances",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Debt"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Debts"
                ],
                "summary": "Créer une dette ou créance",
                "operationId": "4af92119f926413c0796ce5dbe97c7e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "initial_amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Prêt famille"
                                    },
                                    "type": {
                                        "description": "debt = je dois, credit = on me doit",
                                        "type": "string",
                                        "enum": [
                                            "debt",
                                            "credit"
                                        ]
                                    },
                                    "initial_amount": {
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "current_amount": {
                                        "description": "Si différent du montant initial",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "contact_name": {
                                        "type": "string",
                                        "example": "Papa",
                                        "nullable": true
                                    },
                                    "contact_phone": {
                                        "type": "string",
                                        "example": "+241 XX XX XX XX",
                                        "nullable": true
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#EF4444"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dette/créance créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Debt"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/debts/{id}": {
            "get": {
                "tags": [
                    "Debts"
                ],
                "summary": "Détail d'une dette/créance",
                "operationId": "f36f14d7d18d99d9337b2acaa72da853",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Debt"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Debts"
                ],
                "summary": "Modifier une dette/créance",
                "operationId": "d669278483c8af58bf1270acbb5d62cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "debt",
                                            "credit"
                                        ]
                                    },
                                    "initial_amount": {
                                        "type": "integer"
                                    },
                                    "current_amount": {
                                        "type": "integer"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "contact_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "contact_phone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "paid",
                                            "cancelled"
                                        ]
                                    },
                                    "color": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Debt"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Debts"
                ],
                "summary": "Supprimer une dette/créance",
                "operationId": "3788c1bdeac9a2e4d9ff8ac91b3c478d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/debts/{id}/payment": {
            "post": {
                "tags": [
                    "Debts"
                ],
                "summary": "Enregistrer un paiement",
                "operationId": "f9c4fcb0162704e53d08680c08bba43a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Montant du paiement en FCFA",
                                        "type": "integer",
                                        "example": 25000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Debt"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/debts-stats": {
            "get": {
                "tags": [
                    "Debts"
                ],
                "summary": "Statistiques des dettes et créances",
                "operationId": "f3bb9d188db527bbea20f0216f3e6871",
                "responses": {
                    "200": {
                        "description": "Statistiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_debt": {
                                            "description": "Total des dettes actives",
                                            "type": "integer"
                                        },
                                        "total_credit": {
                                            "description": "Total des créances actives",
                                            "type": "integer"
                                        },
                                        "active_debts": {
                                            "type": "integer"
                                        },
                                        "active_credits": {
                                            "type": "integer"
                                        },
                                        "overdue_count": {
                                            "type": "integer"
                                        },
                                        "net_position": {
                                            "description": "créances - dettes",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Debt": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Prêt famille"
                    },
                    "type": {
                        "description": "debt = je dois, credit = on me doit",
                        "type": "string",
                        "enum": [
                            "debt",
                            "credit"
                        ]
                    },
                    "initial_amount": {
                        "type": "integer",
                        "example": 100000
                    },
                    "current_amount": {
                        "type": "integer",
                        "example": 75000
                    },
                    "paid_amount": {
                        "type": "integer",
                        "example": 25000
                    },
                    "remaining_amount": {
                        "type": "integer",
                        "example": 75000
                    },
                    "progress_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 25
                    },
                    "formatted_initial_amount": {
                        "type": "string",
                        "example": "100 000 FCFA"
                    },
                    "formatted_current_amount": {
                        "type": "string",
                        "example": "75 000 FCFA"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "days_until_due": {
                        "type": "integer",
                        "nullable": true
                    },
                    "is_overdue": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "contact_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "contact_phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "paid",
                            "cancelled"
                        ]
                    },
                    "color": {
                        "type": "string",
                        "example": "#EF4444"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via token Sanctum. Utilisez le token reçu lors du login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}